requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.2,<3.0
python-dotenv>=1.0,<2.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.3,<5.0